extract scroll code from renderer into hook
